requirements.txt
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
requests==2.26.0
beautifulsoup4==4.9.3
tensorflow==2.6.0
feature-engine==1.1.0
tld==0.12.6
python-whois==0.7.3
selenium==3.141.0
fastapi==0.68.0
uvicorn==0.14.0
python-multipart==0.0.5